Refactor favor comparison. Remove try catch blocks in util methods
