+++ app/src/main/java/ch/epfl/favo/favor/Favor.java
@@ -225,24 +225,7 @@ public class Favor implements Parcelable, Document {
     this.statusId = other.getStatusId();
   }
   // Overriding equals() to compare two Complex objects
-  @Override
-  public boolean equals(Object o) {
-
-    // If the object is compared with itself then return true
-    if (o == this) {
-      return true;
-    }
-
-    /* Check if o is an instance of Complex or not
-    "null instanceof [type]" also returns false */
-    if (!(o instanceof Favor)) {
-      return false;
-    }
-
-    // typecast o to Complex so that we can compare data members
-    Favor other = (Favor) o;
-
-    // Compare the data members and return accordingly
+  public boolean contentEquals(Favor other) {
     return this.title.equals(other.title)
         && this.description.equals(other.description)
         && this.statusId.equals(other.getStatusId())
